+++ packages/Checkbox/stories/Checkbox.stories.js
@@ -1,16 +1,27 @@
 import React from "react";
 import { storiesOf } from "@storybook/react";
 import { withKnobs } from "@storybook/addon-knobs";
-import { CheckboxStory } from "./Checkbox.stories.styles";
-import CheckboxExample from "./examples/CheckboxExample";
+import L10n from "@paprika/l10n";
+import { Story } from "storybook/assets/styles/common.styles";
+import CheckboxAlign from "./examples/CheckboxAlign";
+import CheckboxGrouping from "./examples/CheckboxGrouping";
 import ShowcaseStory from "./examples/Showcase";
 
 storiesOf("Checkbox", module)
   .addDecorator(withKnobs)
-  .add("Showcase", ShowcaseStory);
+  .add("Showcase", ShowcaseStory)
 
-storiesOf("Checkbox/Automation Tests/Screener", module).add("CheckboxExample", () => (
-  <CheckboxStory>
-    <CheckboxExample size="medium" />
-  </CheckboxStory>
-));
+  .add("Checkbox group", () => (
+    <L10n locale="zh">
+      <Story>
+        <CheckboxGrouping />
+      </Story>
+    </L10n>
+  ))
+  .add("Checkbox Align", () => (
+    <L10n locale="zh">
+      <Story>
+        <CheckboxAlign />
+      </Story>
+    </L10n>
+  ));
